requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
langchain-community>=0.0.10
groq>=0.23.1
python-dotenv>=1.0.0